refactor: Update post comment form to include reply functionality
The code changes in this commit update the post comment form to include the ability to reply to a comment. The form now checks if the "isReply" flag is set and calls the "replyComment" method instead of the "postComment" method. This enhancement improves the user experience by allowing users to easily reply to comments without submitting a new comment.