refactor: Add delete functionality to comment component
The code changes in this commit add the ability to delete a comment in the comment component. When the "Delete" button is clicked, the deleteComment() method is called, which sends a request to the server to delete the comment. This enhancement improves the user experience by allowing authenticated users to easily delete their own comments.