refactor: Update button component to support different button colors
The code changes in this commit update the button component to support different button colors. The component now accepts an input property "color" of type TButtonColor, which can be set to 'primary', 'secondary', or 'danger'. The button's background color is dynamically set based on the value of the "color" property. This enhancement improves the visual appeal and flexibility of the button component.